+++ src/components/jsonrpc/workflows/TGetTransition.php
@@ -1,7 +1,7 @@
 <?php
 namespace extas\components\jsonrpc\workflows;
 
-use extas\components\workflows\exceptions\transitions\ExceptionTransitionMissed;
+use extas\components\exceptions\MissedOrUnknown;
 use extas\interfaces\repositories\IRepository;
 use extas\interfaces\workflows\transitions\ITransition;
 
@@ -19,14 +19,14 @@ trait TGetTransition
      * @param array $query
      * @param string $transitionName
      * @return ITransition|null
-     * @throws
+     * @throws MissedOrUnknown
      */
     protected function getTransition(array $query, string $transitionName): ?ITransition
     {
         $transition = $this->workflowTransitions()->one($query);
 
         if (!$transition) {
-            throw new ExceptionTransitionMissed($transitionName);
+            throw new MissedOrUnknown('transition "' . $transitionName . '"');
         }
 
         return $transition;