+++ tests/jsonrpc/states/StateLoadTest.php
@@ -45,33 +45,25 @@ class StateLoadTest extends TestCase
      */
     public function testValid()
     {
-        $operation = new StateLoad([
-            StateLoad::FIELD__PSR_REQUEST => $this->getPsrRequest('.state.load'),
-            StateLoad::FIELD__PSR_RESPONSE => $this->getPsrResponse()
-        ]);
+        $operation = new StateLoad();
 
         $this->getMagicClass('workflowStates')->create(new State([
             State::FIELD__NAME => 'test2',
             State::FIELD__TITLE => 'test'
         ]));
 
-        $response = $operation();
+        $result = $operation([
+            StateLoad::FIELD__PSR_REQUEST => $this->getPsrRequest('.state.load'),
+            StateLoad::FIELD__PSR_RESPONSE => $this->getPsrResponse()
+        ]);
 
-        /**
-         * @var $jsonRpcResponse IResponse
-         */
-        $jsonRpcResponse = $this->getJsonRpcResponse($response);
-        $this->assertFalse(isset($jsonRpcResponse[IResponse::RESPONSE__ERROR]));
         $this->assertEquals(
             [
-                IResponse::RESPONSE__ID => '2f5d0719-5b82-4280-9b3b-10f23aff226b',
-                IResponse::RESPONSE__VERSION => IResponse::VERSION_CURRENT,
-                IResponse::RESPONSE__RESULT => [
-                    'created_count' => 1,
-                    'got_count' => 2
-                ]
+                'created_count' => 1,
+                'got_count' => 2
             ],
-            $jsonRpcResponse
+            $result,
+            'Incorrect result: ' . print_r($result, true)
         );
     }
 }