+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 6.0.0
+
+- `extas-jsonrpc` dep removed.
+- Use `extas-api-jsonrpc`.
+
 # 5.0.0
 
 - Update to `extas workflow` `4.*`.